==> 99_check_survival.Rmd <==
---
title: "Check survival data"
date: "`r format(Sys.time(), '%d %B %Y')`"
output: word_document
---

Checking the survival data generated by R against the original data.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning=FALSE, message=FALSE, error=FALSE, comment='', dpi=400, dev = "ragg_png")
load('data/FullData.RData') # from 0_read_redcap.R
library(flextable)
library(dplyr)
library(TeachingDemos)
char2seed('sutton')

# function to make table of random selection
selection = function(indata, Hospital, Outcome, n_sample = 3){
f1 = filter(indata, hospital==Hospital, outcome==Outcome, event=='Prior', redcap_version ==3) %>% sample_n(n_sample)
f2 = filter(indata, hospital==Hospital, outcome==Outcome, event=='Yes', redcap_version ==3) %>% sample_n(n_sample)
f3 = filter(indata, hospital==Hospital, outcome==Outcome, event %in% c('Discharge','No'), redcap_version ==3) %>% sample_n(n_sample)
f = bind_rows(f1, f2, f3) %>%
  arrange(event, participant_id) %>%
  select(-int_time, -redcap_version, -hospital, -outcome)
tab = flextable(f) %>%
  theme_box() %>%
  colformat_double(j=5, digits=1) %>%
  autofit()
return(tab)
}
```

# TPCH

## Outcome 4 (clinician-led discussion activity)

```{r}
table = selection(indata= survival_data, Hospital = 'TPCH', Outcome = 'care_review')
table
```

## Outcome 5 (Review of care directives documents)

```{r}
table = selection(indata= survival_data, Hospital = 'TPCH', Outcome = 'care_directive')
table
```

## Outcome 6 (Palliative care referral)

```{r}
table = selection(indata= survival_data, Hospital = 'TPCH', Outcome = 'palliative_care_referral')
table
```

# RBWH

## Outcome 4 (clinician-led discussion activity)

```{r}
table = selection(indata= survival_data, Hospital = 'RBWH', Outcome = 'care_review')
table
```

## Outcome 5 (Review of care directives documents)

```{r}
table = selection(indata= survival_data, Hospital = 'RBWH', Outcome = 'care_directive')
table
```

## Outcome 6 (Palliative care referral)

```{r}
table = selection(indata= survival_data, Hospital = 'RBWH', Outcome = 'palliative_care_referral')
table
```

# GCUH

## Outcome 4 (clinician-led discussion activity)

```{r}
table = selection(indata= survival_data, Hospital = 'GCUH', Outcome = 'care_review')
table
```

## Outcome 5 (Review of care directives documents)

```{r}
table = selection(indata= survival_data, Hospital = 'GCUH', Outcome = 'care_directive')
table
```

## Outcome 6 (Palliative care referral)

```{r}
table = selection(indata= survival_data, Hospital = 'GCUH', Outcome = 'palliative_care_referral')
table
```
